memory display

+++ src/memory.c
@@ -24,5 +24,5 @@ void* memset(void *dest, char val, int count)
 
 unsigned int get_memory_size()
 {
-    return (unsigned int)((float)total_available);
+    return total_available;
 }
